Moving ListResponse/AccessReponse into subfolder util.
